+++ P_260729/Plugins/JMPhysicalGrabber/Source/JMPhysicalGrabberTests/JMPhysicalGrabberTests.Build.cs
@@ -1,0 +1,16 @@
+using UnrealBuildTool;
+
+public class JMPhysicalGrabberTests : ModuleRules
+{
+    public JMPhysicalGrabberTests(ReadOnlyTargetRules Target) : base(Target)
+    {
+        PCHUsage = PCHUsageMode.UseExplicitOrSharedPCHs;
+        PrivateDependencyModuleNames.AddRange(new[]
+        {
+            "Core",
+            "CoreUObject",
+            "Engine",
+            "JMPhysicalGrabber"
+        });
+    }
+}